ArchIntelligent V2.0/requirements.txt
llama-index==0.10.11
llama-index-embeddings-openai==0.1.7
llama-index-llms-openai==0.1.10
llama-index-readers-file==0.1.6
llama-index-vector-stores-postgres==0.1.3
llama-index-retrievers-bm25==0.1.3
pypdf==4.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.1.1
sqlalchemy==2.0.25
pgvector==0.2.4
fastapi==0.103.2
flask==2.3.3
